[MLIR][Python] make Py* types public API
